fix: font and Capitalise
